cpu,gpu

+++ schemas/device_model.py
@@ -3,10 +3,10 @@ from typing import Optional
 from pydantic import BaseModel
 
 
-# 定义嵌套的 DeviceStatus 类
+# 定义嵌套的 DeviceStatus 类,
 class DeviceStatus(BaseModel):
     num: Optional[int] = 0
-    present: Optional[int] = 0
+    present: Optional[float] = 0
 
 
 # 定义 GetDeviceReply 类，使用嵌套的 DeviceStatus
